requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
langchain-community
langchain-qdrant
qdrant-client
sentence-transformers
pydantic
PyYAML
python-dotenv
requests